add authentification

+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -4,12 +4,13 @@ namespace App\Http\Controllers\V1;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ResizeImageRequest;
-use App\Http\Requests\UpdateImageManipulationRequest;
 use App\Http\Resources\V1\ImageManipulationResource;
 use App\Models\Album;
 use App\Models\ImageManipulation;
-use GuzzleHttp\Psr7\UploadedFile;
+use Illuminate\Http\Request;
+use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\URL;
 use Illuminate\Support\Str;
 use Intervention\Image\Facades\Image;
 
@@ -17,27 +18,38 @@ class ImageManipulationController extends Controller
 {
     /**
      * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::collection(ImageManipulation::paginate());
+        return ImageManipulationResource::collection(ImageManipulation::where('user_id', $request->user()->id)->paginate());
     }
-
-    public function byAlbum(Album $album)
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection
+     */
+    public function getByAlbum(Request $request, Album $album)
     {
-        $where = [
-            'album_id' => $album->id
-        ];
+        if ($album->user_id != $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
 
-        return ImageManipulationResource::collection(ImageManipulation::where($where)->paginate());
+        return ImageManipulationResource::collection(ImageManipulation::where([
+            'user_id' => $request->user()->id,
+            'album_id' => $album->id
+        ])->paginate());
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Display the specified resource.
+     *
+     * @param \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\Response
      */
     public function resize(ResizeImageRequest $request)
     {
-
         $all = $request->all();
 
         /** @var UploadedFile|string $image */
@@ -46,19 +58,20 @@ class ImageManipulationController extends Controller
         $data = [
             'type' => ImageManipulation::TYPE_RESIZE,
             'data' => json_encode($all),
-            'user_id' => null
+            'user_id' => $request->user()->id
         ];
-
         if (isset($all['album_id'])) {
-            //TODO
+            $album = Album::find($all['album_id']);
+            if ($album->user_id != $request->user()->id){
+                return abort(403, 'Unauthorized');
+            }
+            $data['album_id'] = $all['album_id'];
         }
-
         $dir = 'images/' . Str::random() . '/';
         $absolutePath = public_path($dir);
         if (!File::exists($absolutePath)) {
             File::makeDirectory($absolutePath, 0755, true);
         }
-
         if ($image instanceof UploadedFile) {
             $data['name'] = $image->getClientOriginalName();
             $filename = pathinfo($data['name'], PATHINFO_FILENAME);
@@ -93,23 +106,21 @@ class ImageManipulationController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     */
-    public function show(ImageManipulation $image)
-    {
-        return new ImageManipulationResource($image);
-    }
-
-    /**
      * Remove the specified resource from storage.
+     *
+     * @param \App\Models\ImageManipulation $imageManipulation
+     * @return \Illuminate\Http\Response
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        if ($image->user_id != $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         $image->delete();
         return response('', 204);
     }
 
-    protected function getWidthAndHeight($w, $h, $originalPath): array
+    protected function getWidthAndHeight($w, $h, $originalPath)
     {
         $image = Image::make($originalPath);
         $originalWidth = $image->width();
